requirements.txt
# Core dependencies for the DTC API SDK
requests>=2.28.0
urllib3>=1.26.0
python-dotenv>=1.0.0

# Optional dependencies for enhanced functionality
typing-extensions>=4.0.0; python_version<"3.10" 